Moved out generated method and var in additional file. Optimized some imports
