+++ pages/creating_file_page.py
@@ -2,7 +2,6 @@ import allure
 
 from utils.browser_helper import BrowserHelper
 from utils.locators import CreatingFilePageLocators
-from utils.logger import logger
 from .base_page import BasePage
 from .header_user_page import HeaderUserPage
