+++ pages/guest_main_page.py
@@ -2,7 +2,6 @@ import allure
 
 from utils.browser_helper import BrowserHelper
 from utils.locators import BasePageLocators
-from utils.logger import logger
 from .base_page import BasePage
 
 
@@ -14,9 +13,6 @@ class GuestMainPage(BrowserHelper, BasePage):
         """
         login_link_button = self.browser.find_element(*BasePageLocators.LOGIN_LINK)
         login_link_button.click()
-        # with allure.step("Do a screenshot with click on login link page"):
-        #     allure.attach(self.browser.get_screenshot_as_png(), name=f'Screenshot {datetime.now()}',
-        #                   attachment_type=AttachmentType.PNG)
 
     @allure.step("Verification of link to login page")
     def should_be_login_link(self):
