+++ pages/settings_repository_page.py
@@ -2,9 +2,9 @@ import allure
 
 from utils.browser_helper import BrowserHelper
 from utils.locators import SettingsRepositoryPageLocators
-from utils.logger import logger
 from .base_page import BasePage
 from .header_user_page import HeaderUserPage
+from utils.generated_data import GeneratedData
 
 
 class SettingsRepositoryPage(HeaderUserPage, BrowserHelper, BasePage):
@@ -23,6 +23,6 @@ class SettingsRepositoryPage(HeaderUserPage, BrowserHelper, BasePage):
 
     @allure.step("Change to new repository name in form")
     def change_repository_name_in_form_and_confirm(self):
-        input_name = self.generate_new_name()
+        input_name = GeneratedData.generate_new_name()
         self.browser.find_element(*SettingsRepositoryPageLocators.REPOSITORY_NAME_FORM).send_keys(input_name)
         self.click_on_element_when_it_become_clickable(*SettingsRepositoryPageLocators.RENAME_CONFIRM_BUTTON)
